requirements.txt
altair==4.2.2
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
Bottleneck==1.3.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.5
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
flit_core==3.6.0
fonttools==4.25.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==6.0.0
importlib-resources==5.2.0
ipykernel==6.19.2
ipython==8.10.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.2.0
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
kiwisolver==1.4.4
lazypredict==0.2.12
lightgbm==3.3.5
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.0
requests==2.28.2
rich==13.3.2
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.19.0
streamlit-lottie==0.0.3
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.38.4
xgboost==1.7.4
zipp==3.11.0
